\name{alterForMedian}
\alias{alterForMedian}
\title{
  Insert a Median Line
}
\description{
  Adds a median line to the DF table.
}
\usage{
alterForMedian(DF, a)
}
\arguments{
  \item{DF}{
  the DF table created by create_DF_rank.
  }
  \item{a}{
  the attributes list, which contains the user specified choice of whether to
  include a median row.
  }
}
\value{
  Returns the DF table.
}
\author{
  Quinn Payton \email{Payton.Quinn@epa.gov}
}
