\name{edPov}
\alias{edPov}
\docType{data}
\title{Example Dataset: Education and Poverty Levels}
\description{
  A datset comparing education and poverty levels among the states.
}
\usage{data(edPov)}
\format{
  A data frame with 51 rows and the following 5 columns:
  \describe{
    \item{state}{full state name.}
    \item{ed}{education rate.}
    \item{pov}{poverty rate.}
    \item{region}{major us region.}
    \item{StateAb}{abbreviated state name.}
  }
}
\note{
  This is just an example data set.
}
\examples{
data(edPov)
head(edPov)
}
\keyword{datasets}


