% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micromap-package.R
\name{internal}
\alias{internal}
\alias{all_atts}
\alias{all_attsb}
\alias{all_equal2}
\alias{axis_opts}
\alias{CatMaps}
\alias{create_DF_rank}
\alias{create_DF_cat}
\alias{dot_legend_att}
\alias{dot_legend_build}
\alias{graph_opts}
\alias{list_att}
\alias{make.string}
\alias{plot_opts}
\alias{RankMaps}
\alias{subplot}
\title{Internal lmplot Functions}
\description{
These functions are internal lmplot functions.
}
\details{
These functions are not to be called by the user. Includes all_atts, all_attsb, all_equal2, axis_opts, CatMaps, create_DF_rank, create_DF_cat, dot_legend_att, dot_legend_build, graph_opts, list_att, make.string, plot_opts, RankMaps, subplot
}
\keyword{internal}
