% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micromap-package.R
\docType{data}
\name{OrEcoLevel3}
\alias{OrEcoLevel3}
\title{Example Dataset: Oregon Level 3 Ecoregion Shapefile}
\format{
The format is: Formal class 'SpatialPointsDataFrame' [package "sp"].
}
\description{
An object of class 'SpatialPolygonsDataFrame' that was created from a
shapefile detailing boundaries of the level 3 ecoregion borders of Oregon.
}
\note{
This is just an example data set.
}
\examples{

data(OrEcoLevel3)
spplot(OrEcoLevel3, zcol="US_L3NAME", col.regions=rainbow(9, s=0.75))

}
\keyword{datasets}
