% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmPlot.r
\name{print.mm}
\alias{print.mm}
\title{Print lmplot Object to a File}
\usage{
\method{print}{mm}(x, name = NULL, res = 300, ...)
}
\arguments{
\item{x}{name of the lmplot object.}

\item{name}{name of the file to be produced. tiff, png, jpeg, and pdf are
recognized.}

\item{res}{resolution in dpi.}

\item{...}{Additional arguments.  (Currently ignored.)}
}
\value{
Returns nothing.
}
\description{
Takes a stored lmplot object of class mm and prints to a file whose type is
implied by the file name.
}
\author{
Quinn Payton \email{Payton.Quinn@epa.gov}
}
