\name{right}
\alias{right}
\title{
  Substring Function
}
\description{
  Mimics the Excel Function of the Same Name
}
\usage{
right(txt, i)
}
\arguments{
  \item{txt}{
  string to be substringed.
  }
  \item{i}{
  number of characters to substring.
  }
}
\value{
  A character value containing the substring.
}
\author{
  Quinn Payton \email{Payton.Quinn@epa.gov}
}
