\name{statesShapefile}
\alias{statesShapefile}
\docType{data}
\title{Example Dataset: States Shapefile}
\description{
	An object of class 'SpatialPolygonsDataFrame' that was created from a
	shapefile of state borders.
}
\usage{data(statesShapefile)}
\format{
  The format is:
Formal class 'SpatialPointsDataFrame' [package "sp"]. 
}
\note{
  This is just an example data set.
}
\examples{
data(statesShapefile)
plot(statesShapefile)
}
\keyword{datasets}


