\name{USstates}
\alias{USstates}
\docType{data}
\title{Example Dataset: U.S. States Polygons}
\description{
	An object of class 'SpatialPolygonsDataFrame' that was created from a
	shapefile of U.S. state borders.
}
\usage{data(USstates)}
\format{
  The format is:
Formal class 'SpatialPolygonsDataFrame' [package "sp"]. 
}
\note{
  This is just an example data set.
}
\examples{
data(USstates)
plot(USstates)
}
\keyword{datasets}


