% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataReorg.r
\name{alterForMedian}
\alias{alterForMedian}
\title{Insert a Median Line}
\usage{
alterForMedian(DF, a)
}
\arguments{
\item{DF}{the DF table created by create_DF_rank.}

\item{a}{the attributes list, which contains the user specified choice of
whether to include a median row.}
}
\value{
Returns the DF table.
}
\description{
Adds a median line to the DF table.
}
\note{
See the Introduction Guide for a full list of the options available
for altering micromaps.
}
\author{
Quinn Payton \email{Payton.Quinn@epa.gov}
}
