% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micromap-package.R
\docType{data}
\name{edPov}
\alias{edPov}
\title{Example Dataset: Education and Poverty Levels}
\format{
A data frame with 51 rows and the following 5 columns: \describe{
\item{state}{full state name.} \item{ed}{education rate.} \item{pov}{poverty
rate.} \item{region}{major us region.} \item{StateAb}{abbreviated state
name.} }
}
\description{
A datset comparing education and poverty levels among the states.
}
\note{
This is just an example data set.
}
\examples{

data(edPov)
head(edPov)

}
\keyword{datasets}
