% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttDefaults.r
\name{labels_att}
\alias{labels_att}
\alias{bar_att}
\alias{bar_cl_att}
\alias{box_summary_att}
\alias{dot_att}
\alias{dot_cl_att}
\alias{map_att}
\alias{ranks_att}
\alias{standard_att}
\title{Default Attribute Lists}
\usage{
labels_att(show = FALSE)
}
\arguments{
\item{show}{If show equals FALSE, then the resulting list is returned
invisibly.}
}
\value{
A list of defaults to be stored in the overall attribute list.
}
\description{
Creates a list of default attribute lists for each panel type.
}
\author{
Quinn Payton \email{Payton.Quinn@epa.gov}
}
