% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.r
\name{right}
\alias{right}
\title{Substring Function}
\usage{
right(txt, i)
}
\arguments{
\item{txt}{string to be substringed.}

\item{i}{number of characters to substring.}
}
\value{
A character value containing the substring.
}
\description{
Mimics the Excel Function of the Same Name
}
\author{
Quinn Payton \email{Payton.Quinn@epa.gov}
}
