% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micromap-package.R
\docType{data}
\name{USstates}
\alias{USstates}
\title{Example Dataset: U.S. States Polygons}
\format{
The format is: Formal class 'SpatialPolygonsDataFrame' [package
"sp"].
}
\description{
An object of class 'SpatialPolygonsDataFrame' that was created from a
shapefile of U.S. state borders.
}
\note{
This is just an example data set.
}
\examples{

data(USstates)
plot(USstates)

}
\keyword{datasets}
